refactor : "PostEntity, UserEntity 경로 및 이름 변경"
홈 도메인 하위의 구조가 부적합하여 post/user 폴더로 이동,
이름을 PostEntity, UserEntity로 변경 및 import 경로 수정

Resolve: #109
