+++ lib/domain/entities/post/post_entity.dart
