+++ lib/domain/entities/user/user_entity.dart
